requirements.txt
neat-python
pygame
psutil>=5.9.0
